Remove ndk debug symbols from Android binary
We initially intended to upload these debug symbols to Sentry, however
we couldn't get Sentry crashes to link to these debug symbols even if we
upload them. This is most probably because if the Rust code panics, the
result is exposed to the native layer as an exception which is the cause
of the crash as far as Sentry is concerned.

Uniffi carries over some of the information from the Rust panic to the
exception, so we still get usual information during a crash. However,
the debug symbols don't seem to make any difference and they have a
heavy cost as it increases the artifact size by 5 MBs. So, at least
until we figure out a way to get some value out of packaging the debug
symbols, we are disabling them.