Fix `login_spec_16_invalid_url`
reqwest was recently updated. It's likely that its internal
implementation changes lead to it now reports a different error when
connecting to an non-existing domain. The casting to `ResolveError`
(`as_dns_error`) no longer works.

This commit add a final check to see whether the error is a connecting
error, and report the error as 'NonExistingSiteError' accordingly.